// ui/src/app/interfaces/contact.ts
export interface Contact {
  _id: string;
  no: Number;
  contactName: string;
  salutation: string;
  mobilePhone: string;
  email: string;
  organization: string;
  dob: string;
  leadSrc: string;
  assignedTo: string;
  creator: string;
  address: string;
  description: string;
  createdTime: string;
  updatedTime: string;
}